+++ tests/tasks_controller_tests.py
@@ -1,20 +1,20 @@
 import unittest
 from unittest.mock import patch
 from freezegun import freeze_time
-from wdc.controller.tasks import start_work_task, list_tasks, get_task_info, amend_task, sort_by_time
-from wdc.classes import WdcTask, WdcTaskInfo
+from wdc.controller.tasks import start_work_task, list_tasks, amend_task, sort_by_time
+from wdc.classes import WdcTask
 
 
 class StartWorkdayTaskFixture(unittest.TestCase):
 
     @freeze_time('2019-10-25')
-    @patch('wdc.controller.tasks.WdcTaskStore.add')
-    def test_only_start_given(self, mock_storage_add):
+    @patch('wdc.controller.tasks.WdcTaskStore')
+    def test_only_start_given(self, mock_storage):
         start_work_task('0800', '', [], '', '')
 
-        mock_storage_add.assert_called()
+        mock_storage.return_value.add_and_save.assert_called()
 
-        call_task = mock_storage_add.call_args.args[0]
+        call_task = mock_storage.return_value.add_and_save.call_args.args[0]
 
         self.assertEqual('2019-10-25', call_task.date)
         self.assertEqual('0800', call_task.start)
@@ -22,12 +22,12 @@ class StartWorkdayTaskFixture(unittest.TestCase):
         self.assertEqual('', call_task.tags)
         self.assertEqual('', call_task.description)
 
-    @patch('wdc.controller.tasks.WdcTaskStore.add')
-    def test_all_parameters_given(self, mock_storage_add):
+    @patch('wdc.controller.tasks.WdcTaskStore')
+    def test_all_parameters_given(self, mock_storage):
         start_work_task('0800', '0815', ['t1', 't2'], 'description', '2020-10-25')
 
-        mock_storage_add.assert_called()
-        call_args = mock_storage_add.call_args.args[0]
+        mock_storage.return_value.add_and_save.assert_called()
+        call_args = mock_storage.return_value.add_and_save.call_args.args[0]
 
         self.assertEqual('2020-10-25', call_args.date)
         self.assertEqual('0800', call_args.start)
@@ -35,12 +35,12 @@ class StartWorkdayTaskFixture(unittest.TestCase):
         self.assertEqual('t1,t2', call_args.tags)
         self.assertEqual('description', call_args.description)
 
-    @patch('wdc.controller.tasks.WdcTaskStore.add')
-    def test_no_end_time_given(self, mock_storage_add):
+    @patch('wdc.controller.tasks.WdcTaskStore')
+    def test_no_end_time_given(self, mock_storage):
         start_work_task('0800', '', ['t1', 't2'], 'description', '2020-10-25')
 
-        mock_storage_add.assert_called()
-        call_args = mock_storage_add.call_args.args[0]
+        mock_storage.return_value.add_and_save.assert_called()
+        call_args = mock_storage.return_value.add_and_save.call_args.args[0]
 
         self.assertEqual('2020-10-25', call_args.date)
         self.assertEqual('0800', call_args.start)
@@ -48,12 +48,12 @@ class StartWorkdayTaskFixture(unittest.TestCase):
         self.assertEqual('t1,t2', call_args.tags)
         self.assertEqual('description', call_args.description)
 
-    @patch('wdc.controller.tasks.WdcTaskStore.add')
-    def test_no_tags_given(self, mock_storage_add):
+    @patch('wdc.controller.tasks.WdcTaskStore')
+    def test_no_tags_given(self, mock_storage):
         start_work_task('0800', '0815', [], 'description', '2020-10-25')
 
-        mock_storage_add.assert_called()
-        call_args = mock_storage_add.call_args.args[0]
+        mock_storage.return_value.add_and_save.assert_called()
+        call_args = mock_storage.return_value.add_and_save.call_args.args[0]
 
         self.assertEqual('2020-10-25', call_args.date)
         self.assertEqual('0800', call_args.start)
@@ -61,12 +61,12 @@ class StartWorkdayTaskFixture(unittest.TestCase):
         self.assertEqual('', call_args.tags)
         self.assertEqual('description', call_args.description)
 
-    @patch('wdc.controller.tasks.WdcTaskStore.add')
-    def test_no_description_given(self, mock_storage_add):
+    @patch('wdc.controller.tasks.WdcTaskStore')
+    def test_no_description_given(self, mock_storage):
         start_work_task('0800', '0815', ['t1', 't2'], '', '2020-10-25')
 
-        mock_storage_add.assert_called()
-        call_args = mock_storage_add.call_args.args[0]
+        mock_storage.return_value.add_and_save.assert_called()
+        call_args = mock_storage.return_value.add_and_save.call_args.args[0]
 
         self.assertEqual('2020-10-25', call_args.date)
         self.assertEqual('0800', call_args.start)
@@ -75,12 +75,12 @@ class StartWorkdayTaskFixture(unittest.TestCase):
         self.assertEqual('', call_args.description)
 
     @freeze_time('2019-10-25')
-    @patch('wdc.controller.tasks.WdcTaskStore.add')
-    def test_no_date_then_today(self, mock_storage_add):
+    @patch('wdc.controller.tasks.WdcTaskStore')
+    def test_no_date_then_today(self, mock_storage):
         start_work_task('0800', '0815', ['t1', 't2'], 'description', '')
 
-        mock_storage_add.assert_called()
-        call_args = mock_storage_add.call_args.args[0]
+        mock_storage.return_value.add_and_save.assert_called()
+        call_args = mock_storage.return_value.add_and_save.call_args.args[0]
 
         self.assertEqual('2019-10-25', call_args.date)
         self.assertEqual('0800', call_args.start)
@@ -100,39 +100,12 @@ class StartWorkdayTaskFixture(unittest.TestCase):
 
 class ListWorkTasksFixture(unittest.TestCase):
     def test_invalid_date(self):
-        self.assertRaises(ValueError, list_tasks, '9999-99-99', False)
+        self.assertRaises(ValueError, list_tasks, '9999-99-99')
+        self.assertRaises(ValueError, list_tasks, '')
 
-    @patch('wdc.controller.tasks.read_all_tasks')
-    def test_return_tasks_for_given_day(self, mock_reader):
-        mock_reader.return_value = [
-            WdcTask(
-                id='task1',
-                date='2020-10-25',
-                start='0800',
-                end='0900',
-                tags='t1',
-                description='test_description1',
-                timestamp='11'
-            ),
-            WdcTask(
-                id='task2',
-                date='2020-10-26',
-                start='0800',
-                end='0900',
-                tags='t2',
-                description='test_description2',
-                timestamp='22'
-            )
-        ]
-
-        results = list_tasks('2020-10-25', True)
-
-        self.assertEqual(1, len(results))
-        self.assertEqual('task1', results[0].id)
-
-    @patch('wdc.controller.tasks.read_all_tasks')
-    def test_task_are_sorted(self, mock_reader):
-        mock_reader.return_value = [
+    @patch('wdc.controller.tasks.WdcTaskStore')
+    def test_task_are_sorted(self, mock_store):
+        mock_store.return_value.get.return_value = [
             WdcTask(
                 id='task1',
                 date='2020-10-25',
@@ -153,179 +126,19 @@ class ListWorkTasksFixture(unittest.TestCase):
             )
         ]
 
-        results = list_tasks('2020-10-25', True)
+        results = list_tasks('2020-10-25')
 
         self.assertEqual(2, len(results))
         self.assertEqual('task2', results[0].id)
         self.assertEqual('task1', results[1].id)
 
-    @patch('wdc.controller.tasks.read_all_tasks')
-    def test_filter_duplicate_tasks(self, mock_reader):
-        mock_reader.return_value = [
-            WdcTask(
-                id='task',
-                date='2020-10-25',
-                start='0800',
-                end='0900',
-                tags='t1',
-                description='test_description1',
-                timestamp='11'
-            ),
-            WdcTask(
-                id='task',
-                date='2020-10-25',
-                start='0800',
-                end='1000',
-                tags='t1',
-                description='test_description1',
-                timestamp='22'
-            )
-        ]
-
-        results = list_tasks('2020-10-25', False)
-
-        self.assertEqual(1, len(results))
-        self.assertEqual('1000', results[0].end)
-
-    @patch('wdc.controller.tasks.read_all_tasks')
-    def test_duplicates_on_beginning(self, mock_reader):
-        mock_reader.return_value = [
-            WdcTask(
-                id='task',
-                date='2020-10-25',
-                start='0800',
-                end='0900',
-                tags='t1',
-                description='test_description1',
-                timestamp='11'
-            ),
-            WdcTask(
-                id='task',
-                date='2020-10-25',
-                start='0800',
-                end='1000',
-                tags='t1',
-                description='test_description1',
-                timestamp='22'
-            ),
-            WdcTask(
-                id='task2',
-                date='2020-10-25',
-                start='10000',
-                end='1130',
-                tags='t2',
-                description='test_description1',
-                timestamp='33'
-            )
-        ]
-
-        results = list_tasks('2020-10-25', False)
-
-        self.assertEqual(2, len(results))
-        self.assertEqual('task', results[0].id)
-        self.assertEqual('1000', results[0].end)
-        self.assertEqual('22', results[0].timestamp)
-        self.assertEqual('task2', results[1].id)
-        self.assertEqual('33', results[1].timestamp)
-
-    @patch('wdc.controller.tasks.read_all_tasks')
-    def test_duplicates_on_end(self, mock_reader):
-        mock_reader.return_value = [
-            WdcTask(
-                id='a2254a3',
-                date='2020-10-25',
-                start='0930',
-                end='',
-                tags='home',
-                description='',
-                timestamp='1595423306302'
-            ),
-            WdcTask(
-                id='c411c941',
-                date='2020-10-25',
-                start='1045',
-                end='',
-                tags='home',
-                description='',
-                timestamp='1595423428554'
-            ),
-            WdcTask(
-                id='c411c941',
-                date='2020-10-25',
-                start='1045',
-                end='1730',
-                tags='home',
-                description='',
-                timestamp='1595430367883'
-            )
-        ]
-
-        results = list_tasks('2020-10-25', False)
-
-        self.assertEqual(2, len(results))
-        self.assertEqual('a2254a3', results[0].id)
-        self.assertEqual('c411c941', results[1].id)
-        self.assertEqual('1730', results[1].end)
-
-
-class FindTaskFixture(unittest.TestCase):
-
-    def test_invalid_task_id(self):
-        self.assertIsNone(get_task_info(''))
-
-    @patch('wdc.controller.tasks.find_tasks')
-    def test_no_tasks_found(self, mock_reader):
-        mock_reader.return_value = []
-
-        self.assertIsNone(get_task_info(''))
-
-
-class WdcTaskInfoFixture(unittest.TestCase):
-    def setUp(self) -> None:
-        self._valid_task_info = WdcTaskInfo(
-            [
-                WdcTask(
-                    id='c411c941',
-                    date='2020-10-25',
-                    start='0930',
-                    end='',
-                    tags='home',
-                    description='',
-                    timestamp='1595423306302'
-                ),
-                WdcTask(
-                    id='c411c941',
-                    date='2020-10-25',
-                    start='0930',
-                    end='1000',
-                    tags='home',
-                    description='',
-                    timestamp='1595423428554'
-                ),
-                WdcTask(
-                    id='c411c941',
-                    date='2020-10-25',
-                    start='1045',
-                    end='1730',
-                    tags='home',
-                    description='',
-                    timestamp='1595430367883'
-                )
-            ]
-        )
-
-    def test_current_is_correct(self):
-        current = self._valid_task_info.current
-
-        self.assertEqual('1595430367883', current.timestamp)
-
 
 class AmendTaskFixture(unittest.TestCase):
 
-    @patch('wdc.controller.tasks.get_task_info')
-    @patch('wdc.controller.tasks.write_task')
-    def test_replace_all(self, mock_writer, mock_reader):
-        mock_reader.return_value = WdcTaskInfo([
+    @patch('wdc.controller.tasks.WdcTaskStore')
+    @patch('wdc.controller.tasks.find_stores')
+    def test_replace_all(self, mock_finder, mock_store):
+        mock_store.return_value.get.return_value = [
             WdcTask(
                 id='c411c941',
                 date='2020-10-25',
@@ -335,13 +148,12 @@ class AmendTaskFixture(unittest.TestCase):
                 description='test description',
                 timestamp='1595423306302'
             )
-        ])
+        ]
+        mock_finder.return_value = [mock_store.return_value]
 
         amend_task('c411c941', tags=['t1', 't2'], start='1000', end='1130', message='test message', date='2020-08-18')
 
-        mock_writer.assert_called()
-
-        call_args = mock_writer.call_args.args[0]
+        call_args = mock_store.return_value.add_and_save.call_args.args[0]
 
         self.assertEqual('c411c941', call_args.id)
         self.assertEqual('t1,t2', call_args.tags)
@@ -350,10 +162,10 @@ class AmendTaskFixture(unittest.TestCase):
         self.assertEqual('test message', call_args.description)
         self.assertEqual('2020-08-18', call_args.date)
 
-    @patch('wdc.controller.tasks.get_task_info')
-    @patch('wdc.controller.tasks.write_task')
-    def test_none_replaced(self, mock_writer, mock_reader):
-        mock_reader.return_value = WdcTaskInfo([
+    @patch('wdc.controller.tasks.WdcTaskStore')
+    @patch('wdc.controller.tasks.find_stores')
+    def test_none_replaced(self, mock_finder, mock_store):
+        mock_store.return_value.get.return_value = [
             WdcTask(
                 id='c411c941',
                 date='2020-10-25',
@@ -363,13 +175,12 @@ class AmendTaskFixture(unittest.TestCase):
                 description='test description',
                 timestamp='1595423306302'
             )
-        ])
+        ]
+        mock_finder.return_value = [mock_store.return_value]
 
         amend_task('c411c941', tags=[], start='', end='', message='', date='')
 
-        mock_writer.assert_called()
-
-        call_args = mock_writer.call_args.args[0]
+        call_args = mock_store.return_value.add_and_save.call_args.args[0]
 
         self.assertEqual('c411c941', call_args.id)
         self.assertEqual('home', call_args.tags)